requirements.txt
annotated-types==0.6.0
anyio==4.3.0
av==11.0.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
ctranslate2==4.1.0
exceptiongroup==1.2.1
fastapi==0.110.2
faster-whisper==1.0.1
filelock==3.13.4
flatbuffers==24.3.25
fsspec==2024.3.1
h11==0.14.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
Jinja2==3.1.3
llvmlite==0.42.0
MarkupSafe==2.1.5
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.3
numba==0.59.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
onnxruntime==1.17.3
openai-whisper==20231117
packaging==24.0
protobuf==5.26.1
pydantic==2.7.0
pydantic_core==2.18.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.12
tiktoken==0.6.0
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
triton==2.2.0
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
